added auto conv for StringList, IntList, DoubleList

+++ pyopenms/pxds/StringList.pxd
@@ -4,8 +4,9 @@ from libcpp.vector cimport vector as libcpp_vector
 cdef extern from "<OpenMS/DATASTRUCTURES/StringList.h>" namespace "OpenMS":
 
     cdef cppclass StringList:
+        # wrap-ignore
         StringList() nogil except +
         StringList(StringList) nogil except +
         StringList(libcpp_vector[libcpp_string]) nogil except +
-        int size() nogil except +
+        size_t size() nogil except +
         libcpp_string at(int) nogil except +
